Set time_base for png encoder instantiation. Fixes #62.
Apparently AVCodecContext::time_base must now be set for encoders, like
mentioned in the docs [1].  It does not say however which value should
be put for single-frame encoders, but a quick grep [2] on the lavc PNG
code does not result in a result, so I guess the PNG encoder simply
ignores this value.  However, avcodec_open2() tests if it is set and
will return -1 if it's unset.

[1]: https://ffmpeg.org/doxygen/3.1/structAVCodecContext.html#ab7bfeb9fa5840aac090e2b0bd0ef7589
[2]: grep time_base ~/tmp/ffmpeg-3.1.3/libavcodec/png*